Speedup startup on Cygwin by avoiding costly system() call.
By chance I discovered during profiling the Vim startup that open-browser was in the top 3 of Vim startup time. Analysis showed that this was caused by an unnecessary system('uname') call in the Mac detection; forking a new process is especially slow on Cygwin, due to the architecture mismatch with the underlying Windows subsystem.