add .gitignore
specifically, this is intended to avoid the repo becoming dirty
when the help file doc/tags is autogenerated. I'm using
Pathogen and using open-browser as a git submodule,
when I run `:Helptags` it generates doc/tags. As a note, I've noticed some
of the other vim plugin repos I have as submodules are using a .gitignore
file in this way.