Add allowed_schemes to customize URI detection
Fix #159: smart-search: incorrect protocol detection (vim's s: or g:)

Problem: Some languages (vimscript, lua) use : in their syntax which is
detected by <Plug>(openbrowser-smart-search) as a URI. This results it
pointlessly opening s:URIExtractor instead of searching for
URIExtractor.

Add a configuration to skip the permissive URI detection and be very
conservative instead.

Tested with this configuration:

  " ~/.vim/ftplugin/lua.vim
  let b:openbrowser_allowed_schemes = [
    \     'http',
    \     'https',
    \ ]

The values are used in a pattern, so this configuration works as well:

  " ~/.vim/ftplugin/lua.vim
  let b:openbrowser_allowed_schemes = [
    \     'http(s)\?',
    \ ]

With that configuration, using <Plug>(openbrowser-smart-search) with the
cursor on : on these Lua comments opens the link:
  -- taken from https://leafo.net/guides/setfenv-in-lua52-and-above.html
  -- taken from http://leafo.net/guides/setfenv-in-lua52-and-above.html
  -- taken from ttps://leafo.net/guides/setfenv-in-lua52-and-above.html
  -- taken from ttps://leafo.net/guides/setfenv-in-lua52-and-above.html
This one searches for 'leafo':
  -- taken from tt://leafo.net/guides/setfenv-in-lua52-and-above.html